requirements/base.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-redis==5.4.0
gunicorn==21.2.0
gevent==24.2.1
django-debug-toolbar==4.3.0
pycryptodome==3.20.0
requests==2.31.0
django-storages==1.14.2
boto3==1.34.64
celery==5.3.6
Pillow==10.2.0